Fix delta time in lua script example.
Former-commit-id: 9dce086c20722c366bef8ad7f5424c09a3a84639

+++ examples/lua-scripts/src/main.cpp
@@ -45,13 +45,13 @@ int main(int argc, char** argv)
         ->registerParser<file::LuaScriptParser>("lua")
         ->queue("script/main.lua");
 
-	Signal<Canvas::Ptr, uint, uint>::Slot nextFrame;
+	Signal<Canvas::Ptr, uint, float>::Slot nextFrame;
 	Signal<file::AssetLibrary::Ptr>::Slot loaded = sceneManager->assets()->complete()->connect(
 		[&](file::AssetLibrary::Ptr assets)
 		{
 			loaded = nullptr;
 
-			nextFrame = canvas->enterFrame()->connect([&](Canvas::Ptr, uint, uint)
+			nextFrame = canvas->enterFrame()->connect([&](Canvas::Ptr, uint, float)
 			{
 				nextFrame = nullptr;
 				root->addComponent(sceneManager->assets()->script("script/main.lua"));
